[Bgen] Simplify the NSValue BindAS generated code. (#22257)
As with previous changes, move to use a method group instead of a lambda
with a using block.

We only add the new methods in those platforms that have the specifict
frameworks by using partial class. That way we keep the code cleaner
than with a collection of #if statements.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>