[dotnet] Ensure appropriate IL Strip value for multi-rid builds (#21391)
IsMultiRidBuild tracks if we are executing an inner build for a
multi-rid case. With this approach, if the user specifies a value for
EnableAssemblyILStripping it will still be honored. Otherwise, to ensure
we don't duplicate assemblies when merging the app bundles due to a
strip vs non-strip scenario, IL stripping for the multi-rid build is
disabled by default.

There have been past examples of optimizing in the outer-build:
https://github.com/xamarin/xamarin-macios/pull/12847
But because IL Stripping is contingent on the evaluation of other
RID-specific properties, not sure if that is a feasible approach here.

Fixes https://github.com/xamarin/xamarin-macios/issues/17115

---------

Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>