[RGen] Return the correct NSNumber static method when we have an array. (#22342)
Array types a special because we have to track the array element type
rathern that the actual SpecialType of the array (which is
SpecialType.None).

We add a new property to our data model to track the special type of the
element and we create a helper tuple to keep the switch expression
clean. This could have been done without the helper tuple, but the code
look a lot worse, the compiler is smart enough to remove that extra
tuple.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>