[Bgen] Improve the generated code for SmarEnum BindAs calls. (#22323)
Because we now have a GetValue method that takes a NativeHandle, we can
simplify the code generated when we are dealing with a smart enum in a
BindAs decorated property/method.

The old generated code is:
```csharp
ret = global::CoreAnimation.CACornerCurveExtensions.GetValue (Runtime.GetNSObject<NSString> (global::ObjCRuntime.Messaging.NativeHandle_objc_msgSend (this.Handle, Selector.GetHandle ("cornerCurve")), false)!);
```
While the new one is:
```csharp
ret = global::NaturalLanguage.NLLanguageExtensions.GetValue (global::ObjCRuntime.Messaging.NativeHandle_objc_msgSendSuper (this.SuperHandle, Selector.GetHandle ("language")));
```

The code is simpler, but there is also a small memory imrpovement. In
the old version of the binding, we were not callin the Dispose method in
the NSString. On the other hand, the new code uses the GetValue method
that is disposing the NSString as soon as it leaves the method and
therefore it does not need the GC to take care of it.

PS: There is no need to change the code that creates the wrapper because
we are using method overloading.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>