[Rgen] Split the naming logic form the emitters. (#22354)
We will want to be able to generate code that references to the
generated classes. Rather than having a pattern placed all over the
project, we create a static class that given a string it will return the
name to be used by the emitters.

This, for example, will later be useful when we want to call the
extension methods for a smart enum.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>