[Rgen] Move the DelegateInfo struct as part of the type info. (#22391)
This is done to accommodate the fact that we want to be able to generate
the trampoline code from a TypeInfo because it's the most general struct
that is both used for parameters and return types.

This later will allow use to have factory methods similar to the ones we
have that target the trampoline code generation.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>