[RGen] Generate code for the return type of a trampoline delegate. (#22409)
Based on the return type of a delegate we need to generate code to
convert the native return type of the trampoline to that used by the
managed delegate.

This new function takes the type info of the return type of the delegate
and the name of the aux variable and generates the correct conversion.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>