[Bgen] Ensure that GC.KeepAlive is called in extension methods. (#22433)
This is an interesting change. After working on PR
https://github.com/dotnet/macios/pull/22417 we noticed that we had a
serious problem in how we called extension methods.

To understand the problem we first need to look at the code generation
for an extension method:
```csharp
[BindingImpl (BindingImplOptions.GeneratedCode | BindingImplOptions.Optimizable)]
internal static float _GetValue (ICIHueSaturationValueGradientProtocol This)
{
  return global::ObjCRuntime.Messaging.float_objc_msgSend (This.Handle, Selector.GetHandle ("value"));
}
```

The issue with the above code is that the GC can collect the 'This'
object while is being used by the objc_msgSend method. For this type of
methods to be safe we would need to perform a call to GC.KeepAlive on
This.

Fixing the generator is not as easy as adding the 'GC.KeepAlive (This)'
after the invocation because while that would work with a method that
returns a wrapped object (an object that is a managed representation of
an unmanaged object) it would not in all other case. The changes in the
generator have to be:

1. For all methods that return a value, create a return variable with
the following special cases:
   * Regular methods that return values that are not wrapped.
   * Methods that have been decorated with BindAs.
   * Protocol constructors which are converted in Create methods.
2. Add the GC.KeepAlive call after the obj

With this commit, the generated code now looks like:
```csharp
[BindingImpl (BindingImplOptions.GeneratedCode | BindingImplOptions.Optimizable)]
internal static float _GetValue (ICIHueSaturationValueGradientProtocol This)
{
  float ret;
  ret = global::ObjCRuntime.Messaging.float_objc_msgSend (This.Handle, Selector.GetHandle ("value"));
  GC.KeepAlive (This);
  return ret;
}
```
This change also opens the doors to fix in the future no extension
methods that have parameters that are wrapped objects that do not return
wrapped objects, but that is work for a different PR.

---------

Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>
Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>