[BGen] Fix argument tests for generic types. (#22453)
Generic parameter types are special and in reflection do not show as the
Type that was given as a constrain. This is problematic because the
generator tests for IsWrappedType or IsNativeObject do not work on a
generic type.

This is a patch to get the current bindings working with generic
parameters from generic classes. Once this is landed we will be able to
detect if the types is a wrapped one and we can call GC.KeepAlive when
needed.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>