[devops] Include the macOS tests in the main test pipeline. (#22420)
* Include the macOS tests (build & test) in the main test pipeline.
* Remove the now unused macos-tests pipelines.

The downside is that we'll get the results for the macOS tests a bit later, because
we won't start building the macOS tests until the main build is complete (as opposed
to previously, when we'd build the macOS tests in parallel with the main build).
This shouldn't slow down PR testing though, because the macOS tests isn't the slowest
part (by far).

This simplifies the existing pipelines, and also makes it possible to eventually
create a single GitHub comment that reports the results from all tests (this will
come in a later PR).