[Rgen] Add support for the ForcedType in both the generator and transformer. (#22416)
Provide a similar attribute to the one we find in bgen. We did not
copy/used the same to give use the freedom to make a difference between
the two of them or extend the rgen one if needed.

We have added support to mark the following:

* Properties
* Parameters
* Return values.

Tests have been added for the 3 possible uses cases.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>
Co-authored-by: Rolf Bjarne Kvinge <rolf@xamarin.com>