[xma] Added targets and tasks to copy modified dll and pdb files back to Windows (#22677)
With the new ICorDebug engine for MAUI, we need the debugger to always load the assemblies from Windows. Given that some dll and pdb files changes as part of the remote build, we need a mechanism to copy those modified files back to the Windows OutputPath, so the debugger can always load the right files. Otherwise, the debugger will try to load files that are not the same that are in the running app, so no symbols will be loaded. This also improves the debugger performance by ensuring the files are not loaded remotely but locally

---------

Co-authored-by: Mauro Agnoletti <mauro.agnoletti@gmail.com>