[Rgen] Add a struct that will be use to track a binding in several platforms. (#22628)
Because the transformer needs to compile the APIs for several platforms,
we might get a discrepancy between the platforms and the selectors
present in them. In order to do the correct transformation we need to
keep track of this fact which can later be used to generate the bindings
in a single file.

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>