[RGen] Add method that can calculate the low level parameter for a trampoline. (#22834)
Add a new method that will take a managed delegate parameter and will
generate the needed parameter syntax for the Invoke and delegate in a
trampoline.

Example of conversions:

Action -> IntrPtr
NSObject -> NativeHandle
Pointer -> Pointer
Array -> NativeHandle

The tests take care to verify that the correct conversions is done for
all the known parameter types.

We need this conversion to fix the reviews in
https://github.com/dotnet/macios/pull/22813

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>