[dotnet] Fix dependency tracking for trimming and native linking. Fixes #21223. (#22911)
Store the non-file inputs to the trimmer and native linker in cache files, and if
these inputs change, then rerun the trimmer and native linker. This is implemented
by computing a hash of all these inputs (properties), store the hash to a file, and
overwrite the file if the hash changes. Then in a subsequent target, we force the
trimmer or native linekr to run if the file with the hash value changed.

This is a similar implementation of how the csc task (C# compiler) creates a hash
of the compiler input in order to know when to re-execute the C# compiler.

Also fix a bug in the Codesign task, where the task would always resign an app bundle
if it contained something inside that needed to be signed separately (it wouldn't
consider that the contained something could already be signed).

Fixes https://github.com/dotnet/macios/issues/21223.