[devops] Clear old Xcode settings. (#22971)
Provisionator isn't able to write to the new location yet, so remove both
settings locations just to make sure everything is pristine before asking the
provisionator to install Xcode.

This will hopefully fix an issue where the build picks up an earlier version of the settings plist with the wrong Xcode version.