[UIKit] Make sure to invoke the correct constructor. Fixes #23122. (#23127)
There are three constructors involved here:

1. (UIBarButtonSystemItem systemItem, EventHandler handler)
2. (UIBarButtonSystemItem systemItem, NSObject target, Selector action)
3. (UIBarButtonSystemItem systemItem, UIAction primaryAction, UIMenu
menu)

The first constructor is a manually written convenience constructor, and
needs
to call another constructor. Either the second or the third constructor
works,
but the third constructor is only available on iOS 16+, while the second
is
available on all iOS version.

This means that calling the third constructor is a bad idea, because it
won't
work on anything before iOS 16. Unfortunately that's exactly what we
did.

So change the code to call the second constructor instead.

This is a regression that was introduced in
f4315389ff2f9d777293f1356b1a9edbe358bdb2.

Fixes https://github.com/dotnet/macios/issues/23122.