[RGen] Generate the StrongDelegates property for a WeakDelegate. (#23232)
This commit allows to generate the strong delegate for a weak delegate
in a bindings. The API provides two ways to configure the strong
property:

1. StrongDelegateType: provide a Type for the new property.
2. StrongDelegateName: Allow to choose the generated property name. If
empty the name will be the name of the WeakDelefate with the Weak prefix
removed.

We have some tests updates:

1. Remove the warnings from the tests. Some are due to the API being
experimenta, other due to a rename. We will remove all errors in a
comming PR.
2. Remove the strong delegate property from the test source since it is
now generated.

This completes the work for support properties (so far).

---------

Co-authored-by: GitHub Actions Autoformatter <github-actions-autoformatter@xamarin.com>