[msbuild] Don't warn about repeated files when looking for duplicated resources. (#23250)
When detecting duplicated resources, make sure to ignore repeated files, we
only want to warn about duplicated resources when they come from different
files.

This fixes an issue where we'd warn about (and not add to the app) a resource
if it was added more than once to the build.